refactor(App): extract useSessionHistory composable
Moves the sessions ref + showHistory flag + fetchSessions/toggleHistory
out of App.vue's script body into a cohesive composable. Narrows the
scope of state that was previously reachable by every function in the
component.
